requirements.txt
pandas 
numpy 
scikit-learn 
matplotlib 
seaborn 
mlflow 
joblib 
fastapi 
uvicorn
google-cloud-bigquery
db-dtypes